WebSite/requirements.txt
arrow==0.8.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==6.6
Django==5.0.1
django-cors-headers==4.3.1
django-webpack-loader==3.0.1
djangorestframework==3.14.0
idna==3.6
optional-django==0.3.0
path.py==8.2.1
pillow==10.2.0
psycopg==3.1.18
psycopg2==2.9.9
python-dateutil==2.5.3
pytz==2024.1
PyYAML==3.11
requests==2.31.0
simplejson==3.8.2
six==1.10.0
sqlparse==0.4.4
structlog==16.1.0
tablib==0.11.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
wheel==0.29.0
